refactor: move repository tests to separate file
