+++ lib/adapter/mod.rs
@@ -1,4 +1,5 @@
 pub mod configuration;
 pub mod models;
 pub mod repository;
+pub(super) mod repository_test;
 pub mod schema;
